+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 This project adheres to [Semantic Versioning](http://semver.org/). All notable changes will be documented in this file.
 
-## [Unreleased](https://github.com/OldSneerJaw/borealis-pg-cli/compare/v1.1.0...HEAD)
+## [1.2.0](https://github.com/OldSneerJaw/borealis-pg-cli/compare/v1.1.0...v1.2.0)
 - Support the new secure tunnel connection info config var (`DATABASE_TUNNEL_BPG_CONN_INFO`)
 - Require SSL/TLS for DB connections when using the `borealis-pg:run` command